requirements.txt
alembic==1.4.2
antiorm==1.2.1
azure-common==1.1.25
azure-core==1.6.0
azure-nspkg==3.0.2
azure-storage-blob==12.3.2
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-nspkg==3.1.0
azure-storage-queue==12.1.2
bcrypt==3.1.7
blinker==1.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
db==0.1.1
db-sqlite3==0.0.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
flask-whooshalchemy-redux==0.7.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
msrest==0.6.16
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
pyodbc==4.0.30
python-dateutil==2.8.1
python-editor==1.0.4
python-ext==1.0.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
Whoosh==2.6.0
WTForms==2.3.1
